Add performance benchmark results and update benchmarking scripts
- Introduced new benchmark results for vector search and insert operations in `bench_results.txt`, `bench_v1_3_4_results.txt`, and `benchmark_results_complete.txt`.
- Added comprehensive benchmarking results for LLM and AQL queries in `comprehensive_windows_llm_aql.json` and `core_perf_windows.json`.
- Updated the PowerShell script `run-bench-tag.ps1` to automate the building and execution of benchmarks, ensuring proper environment setup and output handling.

+++ benchmarks/bench_gnn_embeddings.cpp
@@ -75,7 +75,7 @@ public:
             node.setField("label", "Person");
             node.setField("name", "User " + std::to_string(i));
             node.setField("age", static_cast<int>(age_dist(rng)));
-            node.setField("followers", rng() % 10000);
+            node.setField("followers", static_cast<int64_t>(rng() % 10000));
             
             // Note: addVertex is not needed in current PropertyGraphManager API
             // Vertices are created implicitly when edges reference them
@@ -95,7 +95,7 @@ public:
                     edge.setField("_from", node_ids_[i]);
                     edge.setField("_to", node_ids_[target]);
                     edge.setField("_type", "FOLLOWS");
-                    edge.setField("since", 2020 + (rng() % 5));
+                    edge.setField("since", static_cast<int64_t>(2020 + (rng() % 5)));
 
                     auto st = property_graph_->addEdge(edge, "social");
                     if (!st.ok) {
